Move custom layers to different file
